refactor: decompose fix.ts
closes #146
