+++ package.json
@@ -53,8 +53,8 @@
     "tslib": "^2.6.2"
   },
   "devDependencies": {
-    "@types/fs-extra": "^11.0.2",
     "@qiwi/npm-run-all": "^4.1.7",
+    "@types/fs-extra": "^11.0.2",
     "@types/jest": "^29.5.5",
     "@types/node": "^20.6.2",
     "@types/semver": "^7.5.2",
