+++ src/main/ts/fixes/index.ts
@@ -1,0 +1,6 @@
+export {fixBlankFiles} from './fix-blank-files'
+export {fixContents} from './fix-contents'
+export {fixFilenameVar, fixDirnameVar} from './fix-dirname-var'
+export {fixModuleReferences} from './fix-module-ref'
+export {fixSourceMapRef} from './fix-sourcemap-ref'
+export {fixDefaultExport} from './fix-default-export'
